# Minor optimization to permutation/2.