@JanWielemaker 2015-04-22
Interesting.  There is a price to pay though if (part of) the permutated list is already known. That is not a very unlikely scenario.  In the scenario where your version is faster, you probably want to test each permutation for some condition and that operation will probably outweigh the cost of generating the permutation with a big margin.

Your version

```{prolog}
4 ?- numlist(1,10,L), random_permutation(L,L2), time(permutation(L,L2)).
% 1,104,975 inferences, 0.132 CPU in 0.132 seconds (100% CPU, 8399326 Lips)
L = [1, 2, 3, 4, 5, 6, 7, 8, 9|...],
L2 = [4, 1, 2, 6, 3, 9, 7, 10, 8|...] .
```

The old version

```{prolog}
3 ?- numlist(1,10,L), random_permutation(L,L2), time(permutation(L,L2)).
% 55 inferences, 0.000 CPU in 0.000 seconds (96% CPU, 2670940 Lips)
L = [1, 2, 3, 4, 5, 6, 7, 8, 9|...],
L2 = [5, 9, 10, 3, 8, 7, 2, 4, 6|...] .
```