@eazar001 2015-04-22
Hmm. Indeed the test case with the (+,+) mode speaks volumes by itself. I don't know why this wasn't obvious to me at first. At any rate, thank you for your time.